Module Mc2_backend__Dot

module type S = Mc2_backend.Backend_intf.S

Interface for exporting proofs.

module type Arg = sig ... end

module Default : Arg with type atom := Mc2_core.Atom.t and type hyp := Mc2_core.Clause.t and type lemma := Mc2_core.Clause.t and type assumption := Mc2_core.Clause.t

Provides a reasonnable default to instantiate the Make functor, assuming the original printing functions are compatible with DOT html labels.

module Make : functor (A : Arg with type atom := Mc2_core.Atom.t and type hyp := Mc2_core.Clause.t and type lemma := Mc2_core.Clause.t and type assumption := Mc2_core.Clause.t) -> S with type t := Mc2_core.Proof.t

Functor for making a module to export proofs to the DOT format.