Module Mc2_core__.Internal

Solving facilities

exception Unsat

exception UndecidedLit of Mc2_core__.Solver_types.term

exception Out_of_time

exception Out_of_space

type t

The core solver structure

val create : unit -> t

Create a solver

val add_plugin : t -> Mc2_core.Plugin.Factory.t -> Mc2_core.Plugin.t

add_plugin s f creates a new plugin, stores it into s, and returns it.

raises Failure

if all plugin IDs have been allocated

val plugins : t -> Mc2_core.Plugin.t Iter.t

val get_plugin : t -> Mc2_core__.Solver_types.plugin_id -> Mc2_core.Plugin.t

Get the plugin from its ID

val services : t -> Mc2_core.Service.Registry.t

Service registry

val get_service : t -> 'a Mc2_core.Service.Key.t -> 'a option

Obtain a service by its key

val get_service_exn : t -> 'a Mc2_core.Service.Key.t -> 'a

Obtain a service by its key

val actions : t -> Mc2_core__.Solver_types.actions

Actions available to plugins

val iter_terms : t -> Mc2_core__.Solver_types.term Iter.t

Iterate on all terms known to plugins. Used for checking all variables to assign, and for garbage collection.

val solve : ?⁠gc:bool -> ?⁠restarts:bool -> ?⁠time:float -> ?⁠memory:float -> ?⁠progress:bool -> t -> unit

Try and solves the current set of assumptions.

parameter time

limit, in seconds, measured with Sys.time()

parameter memory

number of words in the heap

returns

() if the current set of clauses is satisfiable

raises Unsat

if a toplevel conflict is found

raises Out_of_time

if out of time

raises Out_of_space

if out of space

val assume : t -> ?⁠tag:int -> Mc2_core__.Solver_types.atom list list -> unit

Add the list of clauses to the current set of assumptions. Modifies the sat solver state in place.

val add_term : t -> Mc2_core__.Solver_types.term -> unit

Add a new generalized variable (i.e term) to the solver. This term will be decided on at some point during solving, wether it appears in clauses or not.

val push : t -> unit

Create a decision level for local assumptions.

raises Unsat

if a conflict is detected in the current state.

val pop : t -> unit

Pop a decision level for local assumptions.

val local : t -> Mc2_core__.Solver_types.atom list -> unit

Add local assumptions

parameter assumptions

list of additional local assumptions to make, removed after the callback returns a value

raises Invalid_argument

if no levels were pushed

Proofs and Models

val eval : Mc2_core__.Solver_types.atom -> bool

Returns the valuation of a term in the current state of the sat solver.

raises UndecidedLit

if the literal is not decided

val eval_level : Mc2_core__.Solver_types.atom -> bool * int

Return the current assignement/evaluation of the boolean term, as well as its decision level. If the level is 0, then it is necessary for the atom to have this value; otherwise it is due to choices that can potentially be backtracked.

raises UndecidedLit

if the literal is not decided not calculable

val model : t -> Mc2_core__.Solver_types.assignment_view list

Returns the model found if the term is satisfiable.

val check : t -> (unit, string) Pervasives.result

Check the satisfiability of the current model. Only has meaning if the solver finished proof search and has returned Sat.

type proof = Mc2_core.Proof.t

val unsat_conflict : t -> Mc2_core__.Solver_types.clause option

Returns the unsat clause found at the toplevel, if it exists (i.e if solve has raised Unsat)

Internal data

These functions expose some internal data stored by the solver, as such great care should be taken to ensure not to mess with the values returned.

val trail : t -> Mc2_core__.Solver_types.trail

Returns the current trail. DO NOT MUTATE

val hyps : t -> Mc2_core__.Solver_types.clause Mc2_core__.Vec.t

Returns the vector of assumptions used by the solver. May be slightly different from the clauses assumed because of top-level simplification of clauses. DO NOT MUTATE

val temp : t -> Mc2_core__.Solver_types.clause Mc2_core__.Vec.t

Returns the clauses coreesponding to the local assumptions. All clauses in this vec are assured to be unit clauses. DO NOT MUTATE

val history : t -> Mc2_core__.Solver_types.clause Mc2_core__.Vec.t

Returns the history of learnt clauses, with no guarantees on order. DO NOT MUTATE

val pp_stats : t CCFormat.printer

val clear_progress : unit -> unit

Clear progress bar