Module Mc2_smtlib__Ast.Ty

type t =

| Ty_bool

| Rat

| Atomic of Mc2_core.ID.t * t list

| Arrow of t * t

val bool : t

val rat : t

val const : Mc2_core.ID.t -> t

val arrow : t -> t -> t

val arrow_l : t list -> t -> t

val equal : t -> t -> bool

val compare : t -> t -> int

val hash : t -> int

val pp : t CCFormat.printer

val unfold : t -> t list * t

unfold ty will get the list of arguments, and the return type of any function. An atomic type is just a function with no arguments

Datatypes

type data = {

data_id : Mc2_core.ID.t;

data_cstors : t Mc2_core.ID.Map.t;

}

Mutually recursive datatypes

module Map : CCMap.S with type Map.key = t

Error Handling

val ill_typed : ('a, Format.formatter, unit, 'b) Pervasives.format4 -> 'a